04 auth

+++ src/controllers/notesController.js
@@ -2,17 +2,19 @@ import createHttpError from 'http-errors';
 import { Note } from '../models/note.js';
 
 export const getAllNotes = async (req, res, next) => {
-  const { page = 1, perPage = 10, tag, search} = req.query;
+  const { page = 1, perPage = 10, tag, search } = req.query;
 
   const skip = (page - 1) * perPage;
-  let notesQuery = Note.find();
+  const userId = req.user._id;
+
+  let notesQuery = Note.find({ userId });
 
   if (tag) {
     notesQuery = notesQuery.where('tag').equals(tag);
   }
 
   if (search) {
-    notesQuery = notesQuery.where({ $text: {$search: search} });
+    notesQuery = notesQuery.where({ $text: { $search: search } });
   }
 
   const [totalNotes, notes] = await Promise.all([
@@ -33,7 +35,9 @@ export const getAllNotes = async (req, res, next) => {
 
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const userId = req.user._id;
+
+  const note = await Note.findById({ _id: noteId, userId });
 
   if (!note) {
     next(createHttpError(404, 'Note not found'));
@@ -44,13 +48,16 @@ export const getNoteById = async (req, res, next) => {
 };
 
 export const createNote = async (req, res, next) => {
-  const newNote = await Note.create(req.body);
+  const userId = req.user._id;
+  const newNote = await Note.create({ ...req.body, userId });
+
   res.status(201).json(newNote);
 };
 
 export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const deletedNote = await Note.findByIdAndDelete(noteId);
+  const userId = req.user._id;
+  const deletedNote = await Note.findByIdAndDelete({ _id: noteId, userId });
 
   if (!deletedNote) {
     next(createHttpError(404, 'Note not found'));
@@ -62,7 +69,8 @@ export const deleteNote = async (req, res, next) => {
 
 export const updateNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const updatedNote = await Note.findByIdAndUpdate(noteId, req.body, {
+    const userId = req.user._id;
+  const updatedNote = await Note.findByIdAndUpdate({ _id: noteId, userId }, req.body, {
     new: true,
   });
